Fix compiler warnings in source code
- Fix shadow warnings in execution.hpp (rename lambda capture variables)
- Fix shadow warnings in slots-groups.cpp and slot-group.cpp example
- Fix sign conversion warnings in stress_test.cpp (proper types/casts)
- Fix sign conversion warnings in execution.cpp test (use size_t)
- Fix float comparison warning in coroutine.cpp (use Catch::Approx)
- Fix destructor override warning in qt-async.hpp
- Fix cast warnings in qt-async.cpp (use char array)
- Fix flaky connect_on_thread Qt test (polling wait pattern)
- Suppress -Wswitch-default (conflicts with enum coverage)

+++ test/async/execution.cpp
@@ -148,13 +148,13 @@ TEST_CASE("Signal sender with let_value", "[execution]") {
 TEST_CASE("Signal sender - multiple concurrent waiters", "[execution][threading]") {
     sigslot::signal<int> sig;
     std::atomic<int> completed{0};
-    constexpr int NUM_WAITERS = 5;
+    constexpr size_t NUM_WAITERS = 5;
 
     std::vector<std::thread> waiters;
     std::vector<int> results(NUM_WAITERS, 0);
 
     // Start multiple waiters
-    for (int i = 0; i < NUM_WAITERS; ++i) {
+    for (size_t i = 0; i < NUM_WAITERS; ++i) {
         waiters.emplace_back([&, i]() {
             auto sender = sigslot::async::as_sender(sig);
             auto [value] = stdexec::sync_wait(std::move(sender)).value();
@@ -173,8 +173,8 @@ TEST_CASE("Signal sender - multiple concurrent waiters", "[execution][threading]
     for (auto& t : waiters)
         t.join();
 
-    REQUIRE(completed == NUM_WAITERS);
-    for (int i = 0; i < NUM_WAITERS; ++i)
+    REQUIRE(completed == static_cast<int>(NUM_WAITERS));
+    for (size_t i = 0; i < NUM_WAITERS; ++i)
         REQUIRE(results[i] == 100);
 }
